+++ back/src/main/java/fr/centralesupelec/thuv/repository/CourseRepository.java
@@ -9,6 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import jakarta.validation.constraints.NotBlank;
@@ -23,8 +24,18 @@ public interface CourseRepository extends JpaRepository<Course, Long>, JpaSpecif
     Set<Course> findByCreator(User admin);
     Course getCourseByLink(String courseLink);
 
-    Page<Course> findAllByTitleContainingIgnoreCaseAndStatusIn(
-            @NotBlank String title, List<CourseStatus> status, Pageable pageable
+    @Query("""
+        SELECT c FROM Course c
+        WHERE 
+            (LOWER(c.title) LIKE LOWER(CONCAT('%', :search, '%')) 
+            OR LOWER(c.creator.name) LIKE LOWER(CONCAT('%', :search, '%')) 
+            OR LOWER(c.creator.lastname) LIKE LOWER(CONCAT('%', :search, '%')))
+            AND c.status IN :status
+    """)
+    Page<Course> searchCoursesByTitleOrCreatorName(
+        @NotBlank @Param("search") String search,
+        @Param("status") List<CourseStatus> status,
+        Pageable pageable
     );
 
     @Transactional